requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
click==8.3.0
colorama==0.4.6
fastapi==0.119.0
fastapi-injector==0.9.0
flake8==7.3.0
h11==0.16.0
idna==3.11
injector==0.22.0
isort==7.0.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pycodestyle==2.14.0
pydantic==2.12.2
pydantic_core==2.41.4
pyflakes==3.4.0
python-dotenv==1.1.1
pytokens==0.1.10
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
